link previews: split linkembed.ts into a link_embed plugin directory
The 850-line module held five responsibilities; it now follows the same
layout as todo_list_multistate: an aggregator (link_embed.ts) requiring
separate editing, UI and autodetect modules, with the commands and the
shared constants in a leaf module (link_embed_commands.ts) so the
satellite form/toolbar files can import them without cycles. The
monolithic spec is split along the same seams, with the shared fixtures
extracted to test/link-embed-kit.ts. No behavior change.
